Added TclObject::getFloat() (as an alternative for getDouble())
Tcl internally only support 'double', not 'float', as floating point
type (in the libtcl 'Tcl_Obj' struct). The openMSX settings are modelled
using Tcl variables, so these also only support 'double'. And the same
for parameters in openMSX commands.

Though sometimes 'float' is accurate enough (and requires less storage,
and allows for wider SIMD operations, ...). So then we get the value via
getDouble() but then immediately assign it to a 'float' variable.

This assignment triggers warnings about narrowing conversions. We can
use the 'narrow()' utility to express this conversion is intended (and
we already did this in some places). Though it's still cleaner to
instead add a getFloat() method and only do the narrowing conversion in
that single method.

+++ src/console/OSDImageBasedWidget.cc
@@ -70,12 +70,12 @@ void OSDImageBasedWidget::setProperty(
 		setRGBA(newRGBA);
 	} else if (propName == "-fadePeriod") {
 		updateCurrentFadeValue();
-		fadePeriod = value.getDouble(interp);
+		fadePeriod = value.getFloat(interp);
 	} else if (propName == "-fadeTarget") {
 		updateCurrentFadeValue();
-		fadeTarget = std::clamp(value.getDouble(interp), 0.0, 1.0);
+		fadeTarget = std::clamp(value.getFloat(interp), 0.0f, 1.0f);
 	} else if (propName == "-fadeCurrent") {
-		startFadeValue = std::clamp(value.getDouble(interp), 0.0, 1.0);
+		startFadeValue = std::clamp(value.getFloat(interp), 0.0f, 1.0f);
 		startFadeTime = Timer::getTime();
 	} else {
 		OSDWidget::setProperty(interp, propName, value);
